requirements.txt
streamlit>=1.35.0
google-genai>=1.0.0
plotly>=5.20.0
requests>=2.31.0
pandas>=2.2.0
networkx>=3.2
biopython>=1.83
